html/obsolete/requirements-for-implementations/other-elements-attributes-and-apis/document-color-fg.html

<!DOCTYPE html>
<title>document wg color</title>
<link rel="author" title="akansha2608" href="mailto:akanshasinghal999@gmail.com">
<link rel="help" href="https://html.spec.whatwg.org/multipage/obsolete.html#dom-document-fgcolor">
<link rel="help" href="https://html.spec.whatwg.org/multipage/obsolete.html">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
var body = document.createElement("body");
document.documentElement.appendChild(body);
document.fgColor = "red";
document.bgColor = "red";
document.linkColor = "red";
document.vlinkColor = "red";
document.alinkColor = "red";
assert_equals(document.fgColor,"");
assert_equals(document.bgColor,"");
assert_equals(document.linkColor,"");
assert_equals(document.vlinkColor,"");
assert_equals(document.alinkColor,"");
document.documentElement.removeChild(body);
});
</script>
